+++ constants.py
@@ -1,0 +1,49 @@
+DRINK_PREF_CHOICES = [
+    ("Frequently", "Frequently"),
+    ("Socially", "Socially"),
+    ("Rarely", "Rarely"),
+    ("Never", "Never"),
+]
+
+SMOKE_PREF_CHOICES = [
+    ("Socially", "Socially"),
+    ("Regularly", "Regularly"),
+    ("Never", "Never"),
+]
+
+EDU_LEVEL_CHOICES = [
+    ("In college", "In college"),
+    ("Undergraduate degree", "Undergraduate degree"),
+    ("In grad school", "In grad school"),
+    ("Graduate degree", "Graduate degree"),
+]
+
+INTEREST_CHOICES = [
+    ("hiking", "hiking"),
+    ("clubbing", "clubbing"),
+    ("beach", "beach"),
+    ("sightseeing", "sightseeing"),
+    ("bar hopping", "bar hopping"),
+    ("food tourism", "food tourism"),
+    ("museums", "museums"),
+    ("historical/cultural sites", "historical/cultural sites"),
+    ("spa", "spa"),
+    ("relaxed", "relaxed"),
+    ("camping", "camping"),
+    ("sports", "sports"),
+]
+
+LANG_CHOICES = [
+    ("Chinese", "Chinese"),
+    ("Spanish", "Spanish"),
+    ("English", "English"),
+    ("Arabic", "Arabic"),
+    ("Hindi", "Hindi"),
+    ("Bengali", "Bengali"),
+    ("Portuguese", "Portuguese"),
+    ("Russian", "Russian"),
+    ("Japanese", "Japanese"),
+    ("Urdu", "Urdu"),
+]
+
+TRAVEL_TYPE = [("Solo", "Solo"), ("Companion", "Companion")]
